Add unique employees query to task view results
In order to effectively build the chart and table on the taskView page,
a listing of unique employees is required. A query returning unique
employees (names & id) was added. Since this new query contains names
associated to an employee ID, the names were removed from the
getAssignedEmployees query, leaving employee id.